Add files via upload
It is a Sample Machine Learning model which is a prototype for detecting the severe issue of the world in the present scenario.  
Why it is a sample model because of the lack of survey and lack of information regarding  covid-19 coronavirus I randomly generate the data but the prototype is working well.

I followed these different phases:

1. Business Understanding ( COVID-19 is characterized as a pandemic by WHO and Whole world, its a need)
2. Collect data or Data gathering (but in our case, it is sample randomly generated data)
************ Statistical Approach ********************
3. Data Analysis ( understating features and labels), 
************ Data Analytical Approach ****************
4. Data preparation
5. Data Modeling
************* Results ***********************************
6. Evaluation Phase
************* Deployment ******************************
7. Model Deployment (using Flask web framework)

Technologies Used:
1. Python
2. CSS (or alternative Bootstrap) 
3. HTML
4. Flask (web framework)
 
Conclusion
The sample prototype model is working absolutely fine and good. It is able to predict whether the person is CoVID-19 Positive or negative,  It also displays the probability percentage. 

Thanks